style: rename command.rs and event.rs
