+++ agent_issuance/src/commands.rs
